+++ src/library/DIPS.Mobile.UI/Components/Labels/CheckTruncatedLabel/iOS/CheckTruncatedLabelHandler.cs
@@ -8,4 +8,9 @@ public partial class CheckTruncatedLabelHandler
     {
         return new iOS.MauiLabel(VirtualView as CheckTruncatedLabel);
     }
+
+    public override Size GetDesiredSize(double widthConstraint, double heightConstraint)
+    {
+        return base.GetDesiredSize(widthConstraint, heightConstraint);
+    }
 }